bluecarbon-citizen-app/api-server/requirements.txt
Flask
Flask-CORS
roboflow
supervision
opencv-python
python-dotenv
Werkzeug
gunicorn
pyserial